+++ Cargo.toml
@@ -13,10 +13,16 @@ url = "2.2.2"
 futures = "0.3"
 serde_json = "1.0"
 csv = "1.1"
+fs2 = "0.4"
 
 [dev-dependencies]
 wiremock = "0.5"
-tokio = { version = "1", features = ["full"] }
+clap = "2.33.3"
 reqwest = { version = "0.11", features = ["json"] }
-csv = "1.1"
+scraper = "0.12.0"
+tokio = { version = "1", features = ["full"] }
+url = "2.2.2"
+futures = "0.3"
 serde_json = "1.0"
+csv = "1.1"
+fs2 = "0.4"
