+++ src/scraper/mod.rs
@@ -1,3 +1,4 @@
 pub mod fetcher;
 pub mod parser;
 pub mod navigator;
+
